% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{admm_solver}
\alias{admm_solver}
\title{Solves the  quadratic programming with linear equality 
and box inequality constraints (non-negative x that sum up to one) using ADMM algorithm.}
\usage{
admm_solver(Dmat, dvec, rho, max_iter, tol)
}
\arguments{
\item{Dmat}{: A (n x n) matrix representing  matrix involved in the quadratic term. Must be positive semidefinite.}

\item{dvec}{: A size n vector representing involved in the linear term of}

\item{rho}{: A scalar value specifying the augmented Lagrange multiplier rho for ADMM algorithm}

\item{max_iter}{: A integer indicating maximum number of iterations for ADMM algorithm.}

\item{tol}{: A scalar indicating the relative tolerance to determine the convergence of ADMM algorithm.}
}
\value{
x : A size n vector containing optimal primal parameters from ADMM
}
\description{
Solves the  quadratic programming with linear equality 
and box inequality constraints (non-negative x that sum up to one) using ADMM algorithm.
}
