% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{att}
\alias{att}
\title{The function estimate the treatment effect}
\usage{
att(X, y, unit_exclude, unit_weights, time_weights)
}
\arguments{
\item{X}{- N*T x k data matrix}

\item{y}{- N*T vector of outcomes}

\item{unit_exclude}{- number of unit to exclude from the estimation. To include all units set this to -1}

\item{unit_weights}{- N x 1 matrix of unit weights}

\item{time_weights}{- T + 1 x 1 matrix of time weights}

\item{weight_type}{- type of weight to be computed. Supported options are "SDID" and "Kernel"}
}
\value{
att: Average Treatment Effect
}
\description{
The function estimate the treatment effect
}
