% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimates the treatment effect using Synthetic Difference in Difference method}
\usage{
estimate(df, covariates, treatment_time, treated_units,
  weight_type = "SDID", resampling_procedure = "Jackknife")
}
\arguments{
\item{df}{Balanced panel data with the following columns: unit, time, treatment, Y (unit id, time id, treatment indicator and outcome variable)}

\item{covariates}{A vector of covariate names to be included. df must contain columns with names included in this variable}

\item{treatment_time}{Time when the treatment occured}

\item{treated_units}{A vector of treated units' ids}

\item{weight_type}{: Type of weights to use. Allowed options are "SDID" for synthetic DID weights or "Kernel" for exponential kernel weighting}

\item{resampling_procedure}{: The algorithm used to obtain distribution of treatment effect. Allowed options are "Jackknife" and "Bootstrap"}
}
\value{
A list that includes:
* est : Point estimate of Average Treatment effect on treated (ATT)
* sigma : Standard error of ATT estimate
* att_arr : N0 size array of resampled estimates of ATT using the resampling procedure
* CI_upper : 95 % confidence interval upper bound for the estimate
* CI_lower : 95 % confidence interval lower bound for the estimate
}
\description{
Estimates the treatment effect using Synthetic Difference in Difference method
}
