% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inference}
\alias{inference}
\title{Inference module. Obtains the empirical distribution of ATT estimate}
\usage{
inference(X, Y, Y00, Y01, Y10, dzeta = 10, rho = 1, tol = 1e-12)
}
\arguments{
\item{X}{- N*T x k data matrix}

\item{Y}{- N*T vector of outcomes}

\item{Y00}{- T0 x N0 matrix of control unit outcomes before the treatment}

\item{Y01}{- T-T0 x N - N0 matrix of control unit outcomes after the treatment}

\item{Y10}{- T0 x N - N0 matrix of control unit outcomes after the treatment}

\item{dzeta}{- Ridge penalty parameter}

\item{rho}{- ADMM parameter}

\item{tol}{- ADMM convergence tolerance}
}
\value{
List containing vector atts of estimates
}
\description{
Inference module. Obtains the empirical distribution of ATT estimate
}
