% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepares the panel dataset and outputs object necessary for the estimation}
\usage{
prepare_data(df, covariates, treatment_time, treated_units)
}
\arguments{
\item{df}{Balanced panel data with the following columns: unit, time, treatment, Y (unit id, time id, treatment indicator and outcome variable)}

\item{covariates}{A vector of covariate names to be included. df must contain columns with names included in this variable}

\item{treatment_time}{Time when the treatment occured}

\item{treated_units}{A vector of treated units' ids}
}
\value{
A list that includes:
* Y : N*T x 1 matrix of outcome variables
* X : Data matrix X used in WLS
* Y00 : T0 x N0 matrix of outcomes for control group before the treatment
* Y01 : T-T0 x N0 matrix of outcomes for control group after treatment
* Y10 : T0 x N - N0 matrix of outcomes for treated group before treatment
* n_covariates : number of covariates used in the estimation
}
\description{
Prepares the panel dataset and outputs object necessary for the estimation
}
