% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_panel.R
\name{simulate_panel}
\alias{simulate_panel}
\title{Creates a simulated balanced panel data to evaluate the method}
\usage{
simulate_panel(N, T, treat_effect, treated_units, treat_date)
}
\arguments{
\item{N}{Number of units in the dataset to be created}

\item{T}{Number of units in the dataset to be created}

\item{treat_effect}{the magnitude of treatment effect}

\item{treated_units}{a vector of treated units}

\item{treat_date}{date at which the treatment happens}
}
\value{
A dataframe with the following columns
* unit : Unit id
* time z : Time id
* treatment : Indicator variable for treatment
* Y : Outcome variable
}
\description{
Creates a simulated balanced panel data to evaluate the method
}
