% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unit_weights}
\alias{unit_weights}
\title{The function find optimal unit weights}
\usage{
unit_weights(Y00, Y10, dzeta, rho, tol, weight_type)
}
\arguments{
\item{Y00}{- T0 x N0 matrix of control unit outcomes before the treatment}

\item{Y10}{- T0 x N - N0 matrix of treated unit outcomes before the treatment}

\item{dzeta}{- Ridge penalty parameter}

\item{rho}{- parameter used in ADMM}

\item{tol}{- Convergence tolerance for ADMM algorithm}

\item{weight_type}{- type of weight to be computed. Supported options are "SDID" and "Kernel"}
}
\value{
N x 1 matrix of time weights for the estimation
}
\description{
The function find optimal unit weights
}
