% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inference}
\alias{inference}
\title{Inference module. Obtains the empirical distribution of ATT estimate}
\usage{
inference(X, Y, Y00, Y01, Y10, dzeta = 10, rho = 1, tol = 1e-12,
  weight_type = "SDID", resampling_procedure = "Jackknife")
}
\arguments{
\item{X}{- N*T x k data matrix}

\item{Y00}{- T0 x N0 matrix of control unit outcomes before the treatment}

\item{Y01}{- T-T0 x N - N0 matrix of control unit outcomes after the treatment}

\item{Y10}{- T0 x N - N0 matrix of control unit outcomes after the treatment}

\item{dzeta}{- Ridge penalty parameter}

\item{rho}{- ADMM parameter}

\item{weight_type}{- type of weight to be computed. Supported options are "SDID" and "Kernel"}

\item{resampling_procedure}{- Resampling procedure used to obtain the distribution. "Jackknife" and "Bootstrap" are supported.}

\item{y}{- N*T vector of outcomes}
}
\value{
List containing vector atts of estimates
}
\description{
Inference module. Obtains the empirical distribution of ATT estimate
}
